[UNCAIM-6][David] fix: convert to int

+++ src/config.py
@@ -11,9 +11,9 @@ class Config:
 
     # JWT configuration
     JWT_SECRET_KEY = os.getenv("JWT_SECRET_KEY")
-    JWT_ACCESS_TOKEN_EXPIRES = os.getenv(
-        "JWT_ACCESS_TOKEN_EXPIRES", 15 * 60
+    JWT_ACCESS_TOKEN_EXPIRES = int(
+        os.getenv("JWT_ACCESS_TOKEN_EXPIRES", 15 * 60)
     )  # Defaults to 15 minutes
-    JWT_REFRESH_TOKEN_EXPIRES = os.getenv(
-        "JWT_REFRESH_TOKEN_EXPIRES", 3 * 24 * 60 * 60
+    JWT_REFRESH_TOKEN_EXPIRES = int(
+        os.getenv("JWT_REFRESH_TOKEN_EXPIRES", 3 * 24 * 60 * 60)
     )  # Defaults to 3 days
